Fixed integer field in mongoose schema

+++ Backend/models/delivered.data.model.js
@@ -17,7 +17,7 @@ const deliveredItems = new Schema ([
         },
 
        barcode: {
-           type: Integer
+           type: Int64
        },
     }
 ]);
